perf(reply): avoid eager harness and abort runtime imports
