+++ src/agents/harness/selection.ts
@@ -1,7 +1,6 @@
 /**
  * Selects and invokes native agent harnesses for embedded run attempts.
  */
-import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import {
   createChildDiagnosticTraceContext,
   createDiagnosticTraceContext,
@@ -12,7 +11,6 @@ import {
 import { formatErrorMessage } from "../../infra/errors.js";
 import { claimHeartbeatContextForUserRun } from "../../infra/heartbeat-outcome-store.js";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
-import { resolveProviderRefOwnership } from "../../plugins/providers.js";
 import { resolveAdmittedRunActiveAssertion } from "../admitted-run-context.js";
 import { resolveSessionAgentIds } from "../agent-scope.js";
 import { resolveGroupToolPolicy } from "../agent-tools.policy.js";
@@ -28,7 +26,6 @@ import type {
   EmbeddedRunAttemptParams,
   EmbeddedRunAttemptResult,
 } from "../embedded-agent-runner/run/types.js";
-import { isCliRuntimeAliasForProvider } from "../model-runtime-aliases.js";
 import {
   unwrapModelHeaderSentinelsForProviderEgress,
   unwrapSecretSentinelsForProviderEgress,
@@ -43,49 +40,37 @@ import {
 } from "../tool-policy.js";
 import type { SystemAgentToolOptions } from "../tools/system-agent-tool.js";
 import { copyCoreTtsAttemptResultProvenance } from "../tools/tts-tool-result-provenance.js";
-import { resolveAgentHarnessAutoSelectionHint } from "./auto-selection.js";
-import { resolveAgentHarnessAvailabilityDecision } from "./availability.js";
 import { createOpenClawAgentHarness, isBuiltInOpenClawAgentHarness } from "./builtin-openclaw.js";
 import { selectContextEngineForTranscriptHost } from "./context-engine-logical-turn.js";
 import { drainPendingContextEngineTurnsBeforeRun } from "./context-engine-turn-attempt.js";
-import { AgentHarnessPreflightError, MissingAgentHarnessError } from "./errors.js";
+import { AgentHarnessPreflightError } from "./errors.js";
 import { createAgentHarnessHostCapabilities } from "./host-capability.js";
 import {
   runAgentHarnessLifecycleAttempt,
   runAgentHarnessLifecycleFinalization,
 } from "./lifecycle.js";
 import type { AgentHarnessPolicy } from "./policy.js";
-import { listRegisteredAgentHarnesses, resolveAgentHarnessOwnerPluginId } from "./registry.js";
 import {
-  buildAgentHarnessSupportContext,
-  compareHarnessSupport,
+  buildAgentHarnessSelectionDecision,
+  resolveAgentHarnessSelectionDecision,
+  type AgentHarnessSelectionParams,
+  type AgentHarnessSelectionDecisionParams,
+  type AgentHarnessSelectionCandidate,
+  type AgentHarnessSelectionDecision as AgentHarnessSelectionFact,
+  type AgentHarnessPreparedModelProvider,
+} from "./selection-decision.js";
+import {
   resolveAgentHarnessPreparedAuthSupport,
   resolveAgentHarnessPreparedRouteSupport,
 } from "./support.js";
-import type { AgentHarness, AgentHarnessSupport, AgentHarnessSupportContext } from "./types.js";
+import type { AgentHarness } from "./types.js";
 
 const log = createSubsystemLogger("agents/harness");
 export { resolveAvailableAgentHarnessPolicy } from "./availability.js";
 
-type AgentHarnessSelectionParams = {
-  provider: string;
-  modelId?: string;
-  modelProvider?: AgentHarnessSupportContext["modelProvider"];
-  config?: OpenClawConfig;
-  agentId?: string;
-  sessionKey?: string;
-  agentHarnessId?: string;
-  agentHarnessRuntimeOverride?: string;
+type AgentHarnessSelectionDecision = Omit<AgentHarnessSelectionFact, "harness"> & {
+  harness: AgentHarness;
 };
-
-type AgentHarnessSelectionDecisionParams = AgentHarnessSelectionParams & {
-  /** Finalized route/auth facts must always pass harness support, including persisted pins. */
-  preparedModelProvider?: boolean;
-};
-
-export type AgentHarnessPreparedModelProvider = NonNullable<
-  AgentHarnessSupportContext["modelProvider"]
->;
 
 const PLUGIN_HARNESS_SENDER_DENY_ALL_PROMPT =
   "Tool and file actions are disabled for this sender by chat policy. If asked to edit files or use tools, say this sender is not allowed by policy; do not imply retrying will help.";
@@ -93,40 +78,6 @@ const PLUGIN_HARNESS_GROUP_DENY_ALL_PROMPT =
   "Tool and file actions are disabled for this chat by policy. If asked to edit files or use tools, say this chat is not allowed by policy.";
 const PLUGIN_HARNESS_RUNTIME_DENY_ALL_PROMPT =
   "Tool and file actions are disabled by runtime policy. If asked to edit files or use tools, say tools are disabled by policy.";
-
-type AgentHarnessSelectionCandidate = {
-  id: string;
-  label: string;
-  pluginId?: string;
-  supported?: boolean;
-  priority?: number;
-  reason?: string;
-};
-
-type AgentHarnessSelectionDecision = {
-  harness: AgentHarness;
-  builtIn: boolean;
-  /** Registry-owned identity; absent only for the built-in runtime. */
-  ownerPluginId?: string;
-  policy: AgentHarnessPolicy;
-  selectedHarnessId: string;
-  selectedReason:
-    | "forced_openclaw"
-    | "forced_plugin"
-    // Implicit Codex preference found no registered Codex harness, so OpenClaw handled the run.
-    | "implicit_plugin_unavailable_openclaw"
-    // Implicit Codex preference cannot reproduce the prepared transport, so OpenClaw handled it.
-    | "implicit_plugin_unsupported_openclaw"
-    // The requested plugin declared OpenClaw as a lossless fallback for this prepared request.
-    | "plugin_declared_fallback_openclaw"
-    // Provider-owned CLI runtime aliases have no agent harness plugin counterpart.
-    | "cli_runtime_passthrough_openclaw"
-    // Auto mode chose a registered plugin harness that supports the provider/model.
-    | "auto_plugin"
-    // Auto mode found no supporting plugin harness, so OpenClaw handled the run.
-    | "auto_openclaw";
-  candidates: AgentHarnessSelectionCandidate[];
-};
 
 type PluginHarnessToolPolicyContext = Pick<
   EmbeddedRunAttemptParams,
@@ -171,10 +122,6 @@ type ResolvedPluginHarnessToolPolicies = {
   safeDeniedToolNames: string[];
   toolPolicyRestricted: boolean;
 };
-
-function listPluginAgentHarnesses(): AgentHarness[] {
-  return listRegisteredAgentHarnesses().map((entry) => entry.harness);
-}
 
 export function selectAgentHarness(params: AgentHarnessSelectionParams): AgentHarness {
   return selectAgentHarnessDecision(params).harness;
@@ -225,146 +172,11 @@ export function agentHarnessExposesOpenClawTools(harnessId: string): boolean {
 function selectAgentHarnessDecision(
   params: AgentHarnessSelectionDecisionParams,
 ): AgentHarnessSelectionDecision {
-  // Keep the probed instance: owner validation must reject replacement during supports().
-  const pluginHarnesses = listPluginAgentHarnesses();
-  const availability = resolveAgentHarnessAvailabilityDecision({
-    ...params,
-    resolveProviderOwnership: () =>
-      resolveProviderRefOwnership({
-        provider: params.provider,
-        config: params.config,
-      }),
-  });
-  const policy = availability.policy;
-  // OpenClaw's built-in harness is intentionally not part of the plugin candidate list. Explicit plugin
-  // runtimes fail closed unless the selected plugin declares OpenClaw as a lossless fallback.
-  const openClawHarness = createOpenClawAgentHarness();
-  const runtime = policy.runtime;
-  if (runtime === "openclaw") {
-    const selectedReason =
-      availability.kind === "implicit-unavailable"
-        ? "implicit_plugin_unavailable_openclaw"
-        : availability.kind === "implicit-unsupported"
-          ? "implicit_plugin_unsupported_openclaw"
-          : availability.kind === "declared-fallback"
-            ? "plugin_declared_fallback_openclaw"
-            : "forced_openclaw";
-    return buildSelectionDecision({
-      harness: openClawHarness,
-      policy,
-      selectedReason,
-      candidates: listHarnessCandidates(pluginHarnesses),
-    });
-  }
-  if (runtime !== "auto") {
-    const forced = pluginHarnesses.find((entry) => entry.id === runtime);
-    if (forced) {
-      const support = availability.support;
-      if (!support || support.supported || support.fallbackRuntime === "openclaw") {
-        if (support && !support.supported) {
-          log.info(
-            `agent harness selected requested=${runtime} selected=${forced.id} reason=private_qa_forced_runtime`,
-          );
-        }
-        return buildSelectionDecision({
-          harness: forced,
-          policy,
-          selectedReason: "forced_plugin",
-          candidates: listHarnessCandidates(pluginHarnesses),
-        });
-      }
-      if (isCliRuntimeAliasForProvider({ runtime, provider: params.provider })) {
-        return buildSelectionDecision({
-          harness: openClawHarness,
-          policy: {
-            ...policy,
-            runtime: "openclaw",
-          },
-          selectedReason: "cli_runtime_passthrough_openclaw",
-          candidates: listHarnessCandidates(pluginHarnesses),
-        });
-      }
-      throw new Error(
-        `Requested agent harness "${runtime}" does not support ${formatProviderModel(params)}${
-          support.reason ? ` (${support.reason})` : ""
-        }.`,
-      );
-    }
-    if (
-      isCliRuntimeAliasForProvider({
-        runtime,
-        provider: params.provider,
-        cfg: params.config,
-      })
-    ) {
-      return buildSelectionDecision({
-        harness: openClawHarness,
-        policy: {
-          ...policy,
-          runtime: "openclaw",
-        },
-        selectedReason: "cli_runtime_passthrough_openclaw",
-        candidates: listHarnessCandidates(pluginHarnesses),
-      });
-    }
-    throw new MissingAgentHarnessError(runtime);
-  }
-
-  const hintedCandidates = pluginHarnesses.map((harness) => ({
-    harness,
-    support: resolveAgentHarnessAutoSelectionHint({ harness, provider: params.provider }),
-  }));
-  const candidates = hintedCandidates.some((entry) => entry.support === undefined)
-    ? (() => {
-        const supportContext = buildAgentHarnessSupportContext({
-          provider: params.provider,
-          modelId: params.modelId,
-          modelProvider: params.modelProvider,
-          requestedRuntime: runtime,
-          config: params.config,
-          agentId: params.agentId,
-          sessionKey: params.sessionKey,
-          preparedModelProvider: params.preparedModelProvider,
-          providerOwnership: resolveProviderRefOwnership({
-            provider: params.provider,
-            config: params.config,
-          }),
-        });
-        return hintedCandidates.map(({ harness, support }) => ({
-          harness,
-          support: support ?? harness.supports(supportContext),
-        }));
-      })()
-    : hintedCandidates.map(({ harness, support }) => ({
-        harness,
-        support: support as AgentHarnessSupport,
-      }));
-  const supported = candidates
-    .filter(
-      (
-        entry,
-      ): entry is {
-        harness: AgentHarness;
-        support: AgentHarnessSupport & { supported: true };
-      } => entry.support.supported,
-    )
-    .toSorted(compareHarnessSupport);
-
-  const selected = supported[0]?.harness;
-  if (selected) {
-    return buildSelectionDecision({
-      harness: selected,
-      policy,
-      selectedReason: "auto_plugin",
-      candidates: candidates.map(toSelectionCandidate),
-    });
-  }
-  return buildSelectionDecision({
-    harness: openClawHarness,
-    policy,
-    selectedReason: "auto_openclaw",
-    candidates: candidates.map(toSelectionCandidate),
-  });
+  const selection = resolveAgentHarnessSelectionDecision(params);
+  return {
+    ...selection,
+    harness: selection.builtIn ? createOpenClawAgentHarness() : selection.harness,
+  };
 }
 
 /** Runs the selected harness's fail-closed settled-turn finalization operation. */
@@ -1050,43 +862,18 @@ function policyDeniesAllTools(policy?: { deny?: string[] }): boolean {
   );
 }
 
-function listHarnessCandidates(harnesses: AgentHarness[]): AgentHarnessSelectionCandidate[] {
-  return harnesses.map((harness) => ({
-    id: harness.id,
-    label: harness.label,
-    pluginId: harness.pluginId,
-  }));
-}
-
-function toSelectionCandidate(entry: {
-  harness: AgentHarness;
-  support: AgentHarnessSupport;
-}): AgentHarnessSelectionCandidate {
-  return {
-    id: entry.harness.id,
-    label: entry.harness.label,
-    pluginId: entry.harness.pluginId,
-    supported: entry.support.supported,
-    priority: entry.support.supported ? entry.support.priority : undefined,
-    reason: entry.support.reason,
-  };
-}
-
 function buildSelectionDecision(params: {
   harness: AgentHarness;
   policy: AgentHarnessPolicy;
   selectedReason: AgentHarnessSelectionDecision["selectedReason"];
   candidates: AgentHarnessSelectionCandidate[];
 }): AgentHarnessSelectionDecision {
-  const builtIn = isBuiltInOpenClawAgentHarness(params.harness);
   return {
+    ...buildAgentHarnessSelectionDecision({
+      ...params,
+      harness: isBuiltInOpenClawAgentHarness(params.harness) ? undefined : params.harness,
+    }),
     harness: params.harness,
-    builtIn,
-    ...(!builtIn ? { ownerPluginId: resolveAgentHarnessOwnerPluginId(params.harness) } : {}),
-    policy: params.policy,
-    selectedHarnessId: params.harness.id,
-    selectedReason: params.selectedReason,
-    candidates: params.candidates,
   };
 }
 
@@ -1107,9 +894,5 @@ function logAgentHarnessSelection(
     runtime: selection.policy.runtime,
     candidates: selection.candidates,
   });
-}
-
-function formatProviderModel(params: { provider: string; modelId?: string }): string {
-  return params.modelId ? `${params.provider}/${params.modelId}` : params.provider;
 }
 /* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
