+++ src/auto-reply/reply/abort.test.ts
@@ -21,16 +21,16 @@ import {
 import { getSessionBindingService } from "../../infra/outbound/session-binding-service.js";
 import { createSuiteTempRootTracker } from "../../test-helpers/temp-dir.js";
 import { resolveAbortCutoffFromContext, shouldSkipMessageByAbortCutoff } from "./abort-cutoff.js";
-import { getAbortMemory } from "./abort-primitives.js";
+import { stopSubagentsForRequester } from "./abort-operation.js";
 import {
-  formatAbortReplyText,
+  getAbortMemory,
   isAbortRequestText,
   isAbortTrigger,
   setAbortMemory,
-  stopSubagentsForRequester,
-  tryFastAbortFromMessage,
-} from "./abort.js";
+} from "./abort-primitives.js";
+import { formatAbortReplyText, tryFastAbortFromMessage } from "./abort.js";
 import { enqueueFollowupRun, getFollowupQueueDepth, type FollowupRun } from "./queue.js";
+import { clearFollowupQueue } from "./queue/state.js";
 import { createReplyOperation, replyRunRegistry } from "./reply-run-registry.js";
 import { testing as replyRunRegistryTesting } from "./reply-run-registry.test-support.js";
 import { buildTestCtx } from "./test-ctx.js";
@@ -266,6 +266,7 @@ describe("abort detection", () => {
   afterEach(async () => {
     for (const key of trackedAbortMemoryKeys) {
       setAbortMemory(key, false);
+      clearFollowupQueue(key);
     }
     trackedAbortMemoryKeys.clear();
     vi.restoreAllMocks();
@@ -475,6 +476,27 @@ describe("abort detection", () => {
     });
 
     expect(result.handled).toBe(true);
+  });
+
+  it("resolves owner authorization after loading cancellation runtime", async () => {
+    const sessionKey = "telegram:123";
+    const sessionId = "session-123";
+    const { root, cfg } = await createAbortConfig({
+      sessionIdsByKey: { [sessionKey]: sessionId },
+    });
+    cfg.commands = { ownerAllowFrom: ["telegram:123"] };
+    enqueueQueuedFollowupRun({ root, cfg, sessionId, sessionKey });
+    const pending = runStopCommand({
+      cfg,
+      sessionKey,
+      from: "telegram:123",
+      to: "telegram:123",
+      senderId: "123",
+    });
+    cfg.commands.ownerAllowFrom = ["telegram:other-owner"];
+
+    await expect(pending).resolves.toEqual({ handled: false, aborted: false });
+    expect(getFollowupQueueDepth(sessionKey)).toBe(1);
   });
 
   it("fast-aborts authorized text slash stop commands before they queue", async () => {
