+++ src/auto-reply/reply/commands-session-abort.ts
@@ -8,13 +8,9 @@ import {
   shouldPersistAbortCutoff,
   type AbortCutoff,
 } from "./abort-cutoff.js";
-import {
-  abortSessionRunTargetWithOutcome,
-  formatAbortReplyText,
-  isAbortTrigger,
-  setAbortMemory,
-  stopSubagentsForRequester,
-} from "./abort.js";
+import { abortSessionRunTargetWithOutcome, stopSubagentsForRequester } from "./abort-operation.js";
+import { isAbortTrigger, setAbortMemory } from "./abort-primitives.js";
+import { formatAbortReplyText } from "./abort.js";
 import { rejectUnauthorizedCommand } from "./command-gates.js";
 import {
   persistAbortTargetEntry,
